tier-1: drop a bunch of redundant packages
These are all included in tier-x or tier-0 already.

Though it does drop bsdtar. This isn't really well-known enough to
belong here. It's included in FCOS because we use it in our scripts,
but it's not a requirement for any tier-1 logic AFAIK and I would be
surprised if users expected it there.